build(ixml): remove ixml libraries for Compa
Compile ixml source files direct into the Compa libraries so I do not
need to install and handle ixml shared and static libraries to be used
by the Compa libraries. Ixml libraries are still available when
compiling with 'pupnp'.

+++ Pupnp/ixml/src/node.cpp
@@ -4,7 +4,7 @@
  * All rights reserved.
  * Copyright (c) 2012 France Telecom All rights reserved.
  * Copyright (C) 2022 GPL 3 and higher by Ingo Höft,  <Ingo@Hoeft-online.de>
- * Redistribution only with this Copyright remark. Last modified: 2022-09-05
+ * Redistribution only with this Copyright remark. Last modified: 2022-09-26
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -329,9 +329,9 @@ int ixmlNode_setNodeValue(IXML_Node* nodeptr, const char* newNodeValue) {
 
 unsigned short ixmlNode_getNodeType(IXML_Node* nodeptr) {
     if (nodeptr != NULL) {
-        return nodeptr->nodeType;
+        return static_cast<unsigned short>(nodeptr->nodeType);
     } else {
-        return (unsigned short)eINVALID_NODE;
+        return static_cast<unsigned short>(eINVALID_NODE);
     }
 }
